Adding transformation logic for cutting process. Additionally adding minor functionality to CuttingLogic and RobotMovement.
TransformCoords provides a method for transforming workspace coordinates
(like trajectory-points), so the cutter-robot knows how to move.
It additionally holds all the necessary transformation matrices, needed
for the cutting process.

CuttingLogic should focus only on the cutting algorithm, not on transfor-
mations or RobotMovements.

RobotMovement only focusses on trajectory calculations and making sure
that the robot moves in the wanted way.

Added button in cuttingMenu for starting cutting-process.

Bug fixes:
- corrupted StringToDouble parsing in Constants fixed
- clearing list of measured marker poses for each calibration fixed